added character class and classes
