+++ app/models/character.py
@@ -38,6 +38,8 @@ class Character(db.Model):
     items = db.relationship('Item', secondary='character_inventories', back_populates='characters')
     # Character and campaign relationship
     campaign = db.relationship('Campaign', secondary='campaign_characters', back_populates='characters')
+    # Character and class relationship
+    classes = db.relationship('Class', secondary='character_classes', back_populates='characters')
 
     def to_dict(self):
         return {
@@ -61,5 +63,6 @@ class Character(db.Model):
             'skills': [skill.to_dict() for skill in self.skills],
             'items': [item.to_dict() for item in self.items],
             'campaigns': [campaign.to_dict_simple() for campaign in self.campaign],
+            'classes': [class_.to_dict() for class_ in self.classes],
             'race': self.race.to_dict()
         }
